pyproject.toml
[project]
name = "sticker-hub"
authors = [
    { name = "Edward Tapado"}
]
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=10.0",
    "pyinstaller>=6.19.0",
    "pyside6>=6.6",
    "python-semantic-release>=10.5.3",
    "requests>=2.31",
]
